[FIX] comment 엔티티에 작성자 join 참조 추가
