+++ .gitignore
@@ -25,6 +25,7 @@ bin/
 out/
 !**/src/main/**/out/
 !**/src/test/**/out/
+src/test/resources/application.yml
 
 ### NetBeans ###
 /nbproject/private/
