+++ src/main/kotlin/com/record/zooc/domain/entity/base/ModifiedTimeEntity.kt
@@ -1,20 +1,20 @@
-package com.record.zooc.domain.entity.time
+package com.record.zooc.domain.entity.base
 
 import com.fasterxml.jackson.annotation.JsonFormat
 import jakarta.persistence.Column
 import jakarta.persistence.EntityListeners
 import jakarta.persistence.MappedSuperclass
-import org.springframework.data.annotation.CreatedDate
+import org.springframework.data.annotation.LastModifiedDate
 import org.springframework.data.jpa.domain.support.AuditingEntityListener
 import java.time.LocalDateTime
 
 @MappedSuperclass
 @EntityListeners(AuditingEntityListener::class)
-abstract class CreatedTimeEntity {
-    @CreatedDate
-    @Column(nullable = false, updatable = false)
+abstract class ModifiedTimeEntity : CreatedTimeEntity() {
+
+    @LastModifiedDate
+    @Column(nullable = false)
     @JsonFormat(pattern = "yyyy-MM-dd HH:mm:ss")
-    var createdAt: LocalDateTime = LocalDateTime.now()
+    var modifiedAt: LocalDateTime = LocalDateTime.now()
         protected set
-    // var로 해야지 시간 저장될 때 오류 안남
 }
