+++ src/main/kotlin/com/record/zooc/domain/entity/comment/Emojicomment.kt
@@ -1,15 +1,16 @@
 package com.record.zooc.domain.entity.comment
 
+import com.record.zooc.domain.entity.Account
 import jakarta.persistence.Column
 import jakarta.persistence.DiscriminatorValue
 
 @DiscriminatorValue("Emoji")
 class Emojicomment(
     content: String,
-    writerId: Long,
+    account: Account,
     recordId: Int,
     emoji: Int
-) : Comment(content, writerId, recordId) {
+) : Comment(content, account, recordId) {
     @Column
     var emoji: Int = emoji
         protected set
