+++ src/main/kotlin/com/record/zooc/domain/repository/UserRepository.kt
@@ -1,6 +1,6 @@
-package com.record.zooc.domain.repository
-
-import User
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface UserRepository : JpaRepository<Long, User>
+// package com.record.zooc.domain.repository
+//
+// import com.record.zooc.domain.entity.User
+// import org.springframework.data.jpa.repository.JpaRepository
+//
+// interface UserRepository : JpaRepository<Long, User>
